# projects/datasite/internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"database/sql"
)

type Cinema struct {
	ID   int64
	Name string
}

type MovieLog struct {
	ID                     int64
	MovieID                int64
	Date                   sql.NullInt64
	CinemaID               sql.NullInt64
	Rating                 sql.NullFloat64
	Review                 sql.NullString
	ReviewContainsSpoilers bool
}